+++ src/features/programs/index.ts
@@ -1,10 +1,19 @@
-export { ProgramsView } from "./ProgramsView";
+export { ProgramHeader, type ProgramTab, type ProgramPatient } from "./ProgramTabs";
+export { OverviewView } from "./OverviewView";
+export { BlocksView } from "./BlocksView";
+export { SessionsView } from "./SessionsView";
+export { PlanningView } from "./PlanningView";
 export { ClientProgramView, ClientSessionView } from "./ClientProgramViews";
+export { buildCalendar, calendarRange } from "./constants";
 export {
+  getBlocks,
   getClientProgram,
   getExerciseLibrary,
   getOwnClient,
+  getPerformanceOverview,
   getSessionDetail,
+  getSessionsBetween,
+  type BlockVM,
   type ProgramVM,
   type SessionDetailVM,
 } from "./data";